reorganized frontend constants, topics and utils
